requirements.txt
asgiref==3.2.9
Django==3.0.4
django-crispy-forms==1.9.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
